Added dependency tslib
For some reason, fuse-box decided to ask me for tslib EVEN THO it was bundling and WORKING without it. I got a message called
"uncaught exception: Package not found tslib"
For some weird reason my fusebox was working without the dependency. Aditionally the official documentation doesn't have word about tslib in the fuse file. It mentions it requires tslib but is not used in the config example they use.